Add error handling for comment and post put and delete

+++ backend/routes/comment.js
@@ -48,7 +48,16 @@ commentRouter.post('/', (req, res) => {
 
 // Edit a comment given a particular ID
 commentRouter.put('/:id', (req, res) => {
+  if (!req.user) {
+    return res.status(401).send({ error: 'unauthorized user' });
+  }
+  if (!req.params.id.match(/^[0-9a-fA-F]{24}$/)) {
+    return res.status(400).send({ error: 'invalid post id' });
+  }
   Comment.findById(req.params.id).then((comment) => {
+    if (comment == null) {
+      return res.status(400).send({ error: 'comment does not exist' });
+    }
     if (comment._commenter.toString().localeCompare(req.user.id.toString()) === 0) {
       Comment.findByIdAndUpdate(req.params.id, { text: req.body.text })
         .then((updatedComment) => res.json(updatedComment))
@@ -61,7 +70,16 @@ commentRouter.put('/:id', (req, res) => {
 
 // Delete a comment given a particular ID
 commentRouter.delete('/:id', (req, res) => {
+  if (!req.user) {
+    return res.status(401).send({ error: 'unauthorized user' });
+  }
+  if (!req.params.id.match(/^[0-9a-fA-F]{24}$/)) {
+    return res.status(400).send({ error: 'invalid post id' });
+  }
   Comment.findById(req.params.id).then((comment) => {
+    if (comment == null) {
+      return res.status(400).send({ error: 'comment does not exist' });
+    }
     if (comment._commenter.toString().localeCompare(req.user.id.toString()) === 0) {
       Comment.findByIdAndDelete(req.params.id)
         .then((deletedComment) => res.json(deletedComment))
